refactor: add question list 분리
